Extra details on what reports to take to the endo
I've seen a lot of questions on Looped about what data, what reports, etc should I bring to the endo.  I've added additional data to the "talk to the endo" page as a way to help provide a little more context.  It would be good to add a set up page for Nightscout Reporter as well as Dexcom Clarity.  I'm happy to write the content and add images to create those pages if there's interest.  Feel free to keep, edit, and delete any of the attached as you see fit.